snc-inf-catastral/requirements.txt
cx_Oracle
sqlalchemy
python-dotenv
pysftp
websocket-client
pandas
openpyxl
pygeoj